refactored the instruction add modal to its own component
